* added dbr_bonds program * cell method moved from file dbr_extend to cells.cc

+++ debyer/utils.h
@@ -45,7 +45,7 @@ public:
     double variance() const { return S_ / (n_ - 1); }
     double stddev() const { return sqrt(variance()); }
     double mean() const { return mean_; }
-    double n() const { return n_; }
+    int n() const { return n_; }
 
 private:
     int n_;
